+++ R/oa2df.R
@@ -5,7 +5,8 @@
 #'
 #' @param data List. Output of \code{oa_request}.
 #' @param entity Character. Scholarly entity of the search.
-#' The argument can be one of c("works", "authors", "venues", "institutions", "concepts").
+#' The argument can be one of
+#' c("works", "authors", "institutions", "concepts", "funders", "sources", "publishers", "topics").
 #' @param abstract Logical. If TRUE, the function returns also the abstract of each item.
 #' Ignored if entity is different from "works". Defaults to TRUE.
 #' @param verbose Logical.
@@ -47,7 +48,7 @@
 #'
 #' @export
 oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = NULL, abstract = TRUE, verbose = TRUE) {
-  if (length(data) == 0){
+  if (length(data) == 0) {
     return(NULL)
   }
 
@@ -59,7 +60,7 @@ oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = N
     return(unlist(data))
   }
 
-  if (entity != "snowball"){
+  if (entity != "snowball") {
     ch <- ifelse(is.null(options$select), "id", options$select[[1]])
     if (!is.null(data[[ch]])) {
       data <- list(data)
@@ -70,11 +71,11 @@ oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = N
     works = works2df(data, abstract, verbose),
     authors = authors2df(data, verbose),
     institutions = institutions2df(data, verbose),
-    venues = venues2df(data, verbose),
     concepts = concepts2df(data, verbose),
     funders = funders2df(data, verbose),
     sources = sources2df(data, verbose),
     publishers = publishers2df(data, verbose),
+    topics = topics2df(data, verbose),
     snowball = snowball2df(data)
   )
 }
@@ -135,20 +136,22 @@ oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = N
 #'
 works2df <- function(data, abstract = TRUE, verbose = TRUE,
                      pb = if (verbose) oa_progress(length(data)) else NULL) {
-
   col_order <- c(
-    "id", "display_name", "author", "ab", "publication_date", "relevance_score",
-    "so", "so_id", "host_organization", "issn_l", "url", "pdf_url",
-    "license", "version", "first_page", "last_page", "volume", "issue", "is_oa",
-    "is_oa_anywhere", "oa_status", "oa_url", "any_repository_has_fulltext",
-    "language", "grants", "cited_by_count", "counts_by_year",
-    "publication_year", "cited_by_api_url", "ids", "doi", "type",
-    "referenced_works", "related_works", "is_paratext", "is_retracted", "concepts"
+    "id", "title", "display_name", "author", "abstract", "doi",
+    "publication_date", "publication_year", "relevance_score", "fwci",
+    "cited_by_count", "counts_by_year", "cited_by_api_url", "ids", "type",
+    "is_oa", "is_oa_anywhere", "oa_status", "oa_url",
+    "any_repository_has_fulltext", "source_display_name", "source_id", "issn_l",
+    "landing_page_url", "pdf_url", "license", "version", "referenced_works",
+    "referenced_works_count", "related_works", "concepts", "topics",
+    "is_paratext", "is_retracted", "language", "grants", "apc",
+    "first_page", "last_page", "volume", "issue"
   )
   works_process <- tibble::tribble(
     ~type, ~field,
     "identical", "id",
     "identical", "display_name",
+    "identical", "title",
     "identical", "publication_date",
     "identical", "doi",
     "identical", "type",
@@ -159,29 +162,23 @@ works2df <- function(data, abstract = TRUE, verbose = TRUE,
     "identical", "is_retracted",
     "identical", "relevance_score",
     "identical", "language",
+    "identical", "fwci",
+    "identical", "referenced_works_count",
     "flat", "grants",
     "flat", "referenced_works",
     "flat", "related_works",
     "rbind_df", "counts_by_year",
     "rbind_df", "concepts",
+    "flat", "apc_list",
+    "flat", "apc_paid",
     "flat", "ids"
   )
 
-  venue_cols <- c(
-    url = "landing_page_url",
-    pdf_url = "pdf_url",
-    is_oa = "is_oa",
-    license = "license",
-    version = "version"
-  )
   so_cols <- c(
-    so_id = "id",
-    so = "display_name",
-    issn_l = "issn_l",
-    host_organization = "host_organization_name"
+    source_id = "id",
+    source_display_name = "display_name",
+    issn_l = "issn_l"
   )
-  inst_cols <- c("id", "display_name", "ror", "country_code", "type", "lineage")
-  empty_inst <- empty_list(inst_cols)
 
   n <- length(data)
   list_df <- vector(mode = "list", length = n)
@@ -198,73 +195,42 @@ works2df <- function(data, abstract = TRUE, verbose = TRUE,
       fields$type,
       SIMPLIFY = FALSE
     )
-
-    author <- venue <- ab <- NULL
-
-    if (!is.null(paper$primary_location)) {
-      so_info <- paper$primary_location["source"]
-      so_info <- if (length(so_info[[1]]) == 0) NA else so_info[[1]]
-      venue_info <- replace_w_na(paper$primary_location[venue_cols])
-      venue <- setNames(
-        c(venue_info, so_info[so_cols]),
-        c(names(venue_cols), names(so_cols))
-      )
+    if (!is.null(sim_fields$publication_date)) {
+      sim_fields$publication_date <- as.Date(sim_fields$publication_date)
     }
-
-    # authorships and affilitation
-    if (!is.null(paper$authorships)) {
-      author <- subs_na(
-        lapply(paper$authorships, function(l) {
-          l_inst <- l$institutions
-          inst_idx <- lengths(l_inst) > 0
-          if (length(inst_idx) > 0 && any(inst_idx)) {
-            first_inst <- l_inst[inst_idx][[1]]
-            first_inst$lineage <- paste(first_inst$lineage, collapse = ", ")
-          } else {
-            first_inst <- empty_inst
-          }
-          first_inst <- prepend(first_inst, "institution")
-          aff_raw <- list(au_affiliation_raw = l$raw_affiliation_string[1])
-          l_author <- if (length(l$author) > 0) {
-            prepend(replace_w_na(l$author), "au")
-          } else {
-            empty_list(c("au_id", "au_display_name", "au_orcid"))
-          }
-          c(l_author, l[c("author_position", "is_corresponding")], aff_raw, first_inst)
-        }), "rbind_df"
-      )
-    }
-
-    # Abstract
-    if (!is.null(paper$abstract_inverted_index) && abstract) {
-      ab <- abstract_build(paper$abstract_inverted_index)
-    }
+    author <- process_paper_authors(paper$authorships)
+    ab <- abstract_build(paper$abstract_inverted_index, abstract)
     paper_biblio <- replace_w_na(paper$biblio)
     open_access <- replace_w_na(paper$open_access)
-    if (length(open_access) > 0){
+    if (length(open_access) > 0) {
       names(open_access)[[1]] <- "is_oa_anywhere"
     }
 
-    out_ls <- c(sim_fields, venue, open_access, paper_biblio, list(author = author, ab = ab))
+    so_info <- paper$primary_location
+    venue <- so_info[names(so_info) != "source"]
+    source <- so_info$source
+    if (!is.null(source)){
+      source <- setNames(source[so_cols], names(so_cols))
+    }
+
+    # Process APC
+    apc <- NULL
+    if (any(lengths(paper[c("apc_list", "apc_paid")]) > 0)) {
+      apc_fields <- list(value = NA, currency = NA, value_usd = NA, provenance = NA)
+      apc <- list(rbind.data.frame(
+        c(type = "list", utils::modifyList(apc_fields, as.list(paper$apc_list))),
+        c(type = "paid", utils::modifyList(apc_fields, as.list(paper$apc_paid)))
+      ))
+    }
+    topics <- process_topics(paper, "score")
+    out_ls <- c(sim_fields, venue, source, open_access, paper_biblio,
+                list(author = author, abstract = ab, apc = apc), topics)
     out_ls[sapply(out_ls, is.null)] <- NULL
     list_df[[i]] <- out_ls
   }
 
   out_df <- rbind_oa_ls(list_df)
   out_df[, intersect(col_order, names(out_df))]
-}
-
-abstract_build <- function(ab) {
-  if (is.null(ab)) {
-    return(NA)
-  }
-  w <- rep(names(ab), lengths(ab))
-  ind <- unlist(ab)
-  if (is.null(ind)) {
-    return("")
-  }
-
-  paste(w[order(ind)], collapse = " ", sep = "")
 }
 
 
@@ -289,17 +255,11 @@ abstract_build <- function(ab) {
 #' # University of Naples Federico II is associated to the OpenAlex id I71267560.
 #'
 #'
-#' query_author <- oa_query(
-#'   identifier = NULL,
+#' res <- oa_fetch(
 #'   entity = "authors",
-#'   last_known_institution.id = "I71267560",
-#'   works_count = ">99"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_author,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   last_known_institutions.id = "I71267560",
+#'   works_count = ">700",
+#'   output = "list"
 #' )
 #'
 #' df <- oa2df(res, entity = "authors")
@@ -315,7 +275,6 @@ authors2df <- function(data, verbose = TRUE,
 
   inst_cols <- c("id", "display_name", "ror", "country_code", "type", "lineage")
   empty_inst <- empty_list(inst_cols)
-  empty_inst$affiliations_other <- list(NULL)
 
   author_process <- tibble::tribble(
     ~type, ~field,
@@ -328,7 +287,6 @@ authors2df <- function(data, verbose = TRUE,
     "identical", "relevance_score",
     "flat", "display_name_alternatives",
     "rbind_df", "counts_by_year",
-    "rbind_df", "x_concepts",
     "flat", "ids"
   )
 
@@ -344,27 +302,29 @@ authors2df <- function(data, verbose = TRUE,
       fields$type,
       SIMPLIFY = FALSE
     )
-    sub_affiliation <- item$last_known_institution
-    if (!is.null(sub_affiliation)) {
+
+    # current affiliation
+    sub_affiliation <- item$last_known_institutions
+    if (!is.null(sub_affiliation) && length(sub_affiliation)) {
+      sub_affiliation <- sub_affiliation[[1]]
       if (is.na(sub_affiliation[[1]])) {
         sub_affiliation <- empty_inst
       }
-      if (length(sub_affiliation$lineage) > 1) {
-        sub_affiliation$lineage <- paste(sub_affiliation$lineage, collapse = ", ")
-      }
+      sub_affiliation$lineage <- paste(sub_affiliation$lineage, collapse = ", ")
       sub_affiliation <- prepend(sub_affiliation, "affiliation")
     }
     sub_affiliation <- replace_w_na(sub_affiliation)
 
+    # all affiliations
     if (!is.null(item$affiliations)) {
-      affiliations_other <- sapply(item$affiliations, function(x) x$institution$id)
-      if (!is.null(sub_affiliation$affiliation_id)) {
-        affiliations_other <- affiliations_other[affiliations_other != sub_affiliation$affiliation_id]
-      }
-      sub_affiliation$affiliations_other <- list(affiliations_other)
+      l_inst <- lapply(item$affiliations, function(x) x$institution)
+      affs <- list(affiliations = process_affil(l_inst))
+    } else {
+      affs <- NULL
     }
 
-    list_df[[i]] <- c(sim_fields, sub_affiliation)
+    topics <- process_topics(item, "count")
+    list_df[[i]] <- c(sim_fields, sub_affiliation, affs, topics)
   }
 
   col_order <- c(
@@ -372,8 +332,7 @@ authors2df <- function(data, verbose = TRUE,
     "ids", "orcid", "works_count", "cited_by_count", "counts_by_year",
     "affiliation_display_name", "affiliation_id", "affiliation_ror",
     "affiliation_country_code", "affiliation_type", "affiliation_lineage",
-    "affiliations_other",
-    "x_concepts", "works_api_url"
+    "affiliations", "topics", "works_api_url"
   )
 
   out_df <- rbind_oa_ls(list_df)
@@ -398,16 +357,11 @@ authors2df <- function(data, verbose = TRUE,
 #'
 #' # Query to search information about all Italian educational institutions
 #'
-#' query_inst <- oa_query(
+#' res <- oa_fetch(
 #'   entity = "institutions",
 #'   country_code = "it",
-#'   type = "education"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_inst,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   type = "education",
+#'   output = "list"
 #' )
 #'
 #' oa2df(res, entity = "institutions")
@@ -439,7 +393,6 @@ institutions2df <- function(data, verbose = TRUE,
     "flat", "display_name_acronyms",
     "row_df", "geo",
     "rbind_df", "counts_by_year",
-    "rbind_df", "x_concepts",
     "rbind_df", "associated_institutions",
     "flat", "ids"
   )
@@ -464,7 +417,8 @@ institutions2df <- function(data, verbose = TRUE,
         )
       )
     }
-    list_df[[i]] <- c(sim_fields, interna)
+    topics <- process_topics(item, "count")
+    list_df[[i]] <- c(sim_fields, interna, topics)
   }
 
 
@@ -474,93 +428,7 @@ institutions2df <- function(data, verbose = TRUE,
     "homepage_url", "image_url", "image_thumbnail_url",
     "associated_institutions", "relevance_score", "works_count",
     "cited_by_count", "counts_by_year",
-    "works_api_url", "x_concepts", "updated_date", "created_date"
-  )
-
-  out_df <- rbind_oa_ls(list_df)
-  out_df[, intersect(col_order, names(out_df))]
-}
-
-
-#' Convert OpenAlex collection of venues' records from list format to data frame
-#'
-#' It converts bibliographic collection of venues' records gathered from OpenAlex database \href{https://openalex.org/}{https://openalex.org/} into data frame.
-#' The function converts a list of venues' records obtained using \code{oa_request} into a data frame/tibble.
-#'
-#' @inheritParams works2df
-#'
-#' @return a data.frame.
-#'
-#' For more extensive information about OpenAlex API, please visit: <https://docs.openalex.org>
-#'
-#'
-#' @examples
-#' \dontrun{
-#'
-#' # Query to search information about the Journal of Informetrics (OA id:V205292342)
-#'
-#'
-#' query_inst <- oa_query(
-#'   identifier = "V205292342",
-#'   entity = "venues"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_inst,
-#'   count_only = FALSE,
-#'   verbose = FALSE
-#' )
-#'
-#' df <- oa2df(res, entity = "venues")
-#'
-#' df
-#' }
-#'
-#' @export
-venues2df <- function(data, verbose = TRUE,
-                      pb = if (verbose) oa_progress(length(data)) else NULL) {
-
-  n <- length(data)
-  list_df <- vector(mode = "list", length = n)
-  venue_process <- tibble::tribble(
-    ~type, ~field,
-    "identical", "id",
-    "identical", "display_name",
-    "identical", "host_organization_name",
-    "identical", "works_count",
-    "identical", "cited_by_count",
-    "identical", "is_oa",
-    "identical", "is_in_doaj",
-    "identical", "homepage_url",
-    "identical", "works_api_url",
-    "identical", "type",
-    "identical", "relevance_score",
-    "flat", "issn_l",
-    "flat", "issn",
-    "rbind_df", "counts_by_year",
-    "rbind_df", "x_concepts",
-    "flat", "ids"
-  )
-
-  for (i in seq.int(n)) {
-    if (verbose) pb$tick()
-
-    item <- data[[i]]
-
-    fields <- venue_process[venue_process$field %in% names(item), ]
-    sim_fields <- mapply(
-      function(x, y) subs_na(item[[x]], type = y),
-      fields$field,
-      fields$type,
-      SIMPLIFY = FALSE
-    )
-    list_df[[i]] <- sim_fields
-  }
-
-  col_order <- c(
-    "id", "display_name", "host_organization_name", "issn", "issn_l", "is_oa", "is_in_doaj",
-    "ids", "homepage_url", "relevance_score", "works_count", "cited_by_count",
-    "counts_by_year", "x_concepts", "works_api_url", "type"
+    "works_api_url", "topics", "updated_date", "created_date"
   )
 
   out_df <- rbind_oa_ls(list_df)
@@ -586,15 +454,10 @@ venues2df <- function(data, verbose = TRUE,
 #' # Query to search information about all Italian educational institutions
 #'
 #'
-#' query_inst <- oa_query(
+#' res <- oa_query(
 #'   entity = "concepts",
-#'   display_name.search = "electrodynamics"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_inst,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   display_name.search = "electrodynamics",
+#'   output = "list"
 #' )
 #'
 #' df <- oa2df(res, entity = "concepts")
@@ -605,7 +468,6 @@ venues2df <- function(data, verbose = TRUE,
 #' @export
 concepts2df <- function(data, verbose = TRUE,
                         pb = if (verbose) oa_progress(length(data)) else NULL) {
-
   concept_process <- tibble::tribble(
     ~type, ~field,
     "identical", "id",
@@ -696,7 +558,6 @@ concepts2df <- function(data, verbose = TRUE,
 #' @export
 funders2df <- function(data, verbose = TRUE,
                        pb = if (verbose) oa_progress(length(data)) else NULL) {
-
   funder_process <- tibble::tribble(
     ~type, ~field,
     "identical", "id",
@@ -770,7 +631,6 @@ funders2df <- function(data, verbose = TRUE,
 #' @export
 sources2df <- function(data, verbose = TRUE,
                        pb = if (verbose) oa_progress(length(data)) else NULL) {
-
   source_process <- tibble::tribble(
     ~type, ~field,
     "identical", "id",
@@ -788,14 +648,13 @@ sources2df <- function(data, verbose = TRUE,
     "identical", "is_in_doaj",
     "flat", "ids",
     "identical", "homepage_url",
-    "identical", "apc_prices",
+    "rbind_df", "apc_prices",
     "identical", "apc_usd",
     "identical", "country_code",
     "flat", "societies",
     "flat", "alternate_titles",
     "identical", "abbreviated_title",
     "identical", "type",
-    "rbind_df", "x_concepts",
     "rbind_df", "counts_by_year",
     "identical", "works_api_url",
     "identical", "updated_date",
@@ -816,7 +675,8 @@ sources2df <- function(data, verbose = TRUE,
       fields$type,
       SIMPLIFY = FALSE
     )
-    list_df[[i]] <- sim_fields
+    topics <- process_topics(item, "count")
+    list_df[[i]] <- c(sim_fields, topics)
   }
 
   out_df <- rbind_oa_ls(list_df)
@@ -854,7 +714,6 @@ sources2df <- function(data, verbose = TRUE,
 #' @export
 publishers2df <- function(data, verbose = TRUE,
                           pb = if (verbose) oa_progress(length(data)) else NULL) {
-
   publisher_process <- tibble::tribble(
     ~type, ~field,
     "identical", "id",
@@ -897,6 +756,85 @@ publishers2df <- function(data, verbose = TRUE,
 
   out_df <- rbind_oa_ls(list_df)
   out_df
+}
+
+
+#' Convert OpenAlex collection of topics' records from list format to data frame
+#'
+#' It converts collection of topics' records gathered from the OpenAlex database.
+#' The function converts a list of topics' records obtained using \code{oa_request} into a data frame/tibble.
+#'
+#' @inheritParams works2df
+#'
+#' @return a data.frame.
+#'
+#' For more extensive information about OpenAlex API, please visit: <https://docs.openalex.org>
+#'
+#'
+#' @examples
+#' \dontrun{
+#'
+#' # Query to search information about all Italian educational institutions
+#'
+#'
+#' res <- oa_query(
+#'   entity = "topics",
+#'   display_name.search = "electrodynamics",
+#'   output = "list"
+#' )
+#'
+#' df <- oa2df(res, entity = "topics")
+#'
+#' df
+#' }
+#'
+#' @export
+topics2df <- function(data, verbose = TRUE,
+                        pb = if (verbose) oa_progress(length(data)) else NULL) {
+  topic_process <- tibble::tribble(
+    ~type, ~field,
+    "identical", "id",
+    "identical", "display_name",
+    "identical", "description",
+    "flat", "ids",
+    "identical", "relevance_score",
+    "identical", "works_count",
+    "identical", "cited_by_count",
+    "identical", "updated_date",
+    "identical", "created_date",
+    "rbind_df", "siblings",
+    "flat", "keywords"
+  )
+
+  n <- length(data)
+  list_df <- vector(mode = "list", length = n)
+
+  for (i in seq.int(n)) {
+    if (verbose) pb$tick()
+
+    item <- data[[i]]
+    fields <- topic_process[topic_process$field %in% names(item), ]
+    sim_fields <- mapply(
+      function(x, y) subs_na(item[[x]], type = y),
+      fields$field,
+      fields$type,
+      SIMPLIFY = FALSE
+    )
+    domains <- unlist(item[c("subfield", "field", "domain")], recursive = FALSE)
+    domains <- as.data.frame(do.call(cbind, domains))
+    names(domains) <- gsub("\\.", "_", names(domains))
+    list_df[[i]] <- c(sim_fields, domains)
+  }
+
+  col_order <- c(
+    "id", "display_name", "description", "keywords", "ids",
+    "subfield_id", "subfield_display_name", "field_id", "field_display_name",
+    "domain_id", "domain_display_name", "siblings", "relevance_score",
+    "works_count", "cited_by_count", "updated_date", "created_date"
+  )
+
+  out_df <- rbind_oa_ls(list_df)
+  out_df[, intersect(col_order, names(out_df))]
 }